% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_projects.R
\name{build_pep_resource_targets_prj}
\alias{build_pep_resource_targets_prj}
\title{Target factory to build targets from a PEP}
\usage{
build_pep_resource_targets_prj(p)
}
\arguments{
\item{p}{PEP defining targets to build}
}
\description{
Target factory to build targets from a PEP
}
