% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitar.R
\name{get_tar_dirs}
\alias{get_tar_dirs}
\title{Gets a tar_dirs folder from either argument or R option}
\usage{
get_tar_dirs(tar_dirs)
}
\arguments{
\item{tar_dirs}{Folder or list of folders (or NULL if using option)}
}
\description{
Gets a tar_dirs folder from either argument or R option
}
