% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_projects.R
\name{load_custom}
\alias{load_custom}
\title{Helper function to build targets from function calls with custom args}
\usage{
load_custom(tname, func, args, vals)
}
\arguments{
\item{tname}{Target name}

\item{func}{Function to call}

\item{args}{List of arg names}

\item{vals}{List of arg values}
}
\description{
This function is a target factory that will produce a target for an
arbitrary function call with arbitrary arguments and values.
}
