% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptar.R
\name{peptar_exec}
\alias{peptar_exec}
\title{Execute a function across all target folders specified in the current Project.}
\usage{
peptar_exec(p, ...)
}
\arguments{
\item{p}{A pepr::Project object representing your current project.}

\item{...}{Function and parameters to execute}
}
\description{
Execute a function across all target folders specified in the current Project.
}
