% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptar.R
\name{peptar_folders}
\alias{peptar_folders}
\title{Retrieves tar folders from PEP}
\usage{
peptar_folders(p)
}
\arguments{
\item{p}{A pepr::Project object representing your current project.}
}
\description{
Retrieves tar folders from PEP
}
