% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptar.R
\name{peptar_make}
\alias{peptar_make}
\title{Execute tar_make across all target folders specified in the current Project.}
\usage{
peptar_make(p, ...)
}
\arguments{
\item{p}{A pepr::Project object representing your current project.}

\item{...}{Parameters to pass to tar_make}
}
\description{
Execute tar_make across all target folders specified in the current Project.
}
