% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptar.R
\name{peptar_meta}
\alias{peptar_meta}
\title{Execute tar_meta across all target folders specified in the current Project.}
\usage{
peptar_meta(p, ...)
}
\arguments{
\item{p}{A pepr::Project object representing your current project.}

\item{...}{Parameters to pass to tar_meta}
}
\description{
Execute tar_meta across all target folders specified in the current Project.
}
