% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptar.R
\name{peptar_path}
\alias{peptar_path}
\title{Find path to a target from all your PEP target folders}
\usage{
peptar_path(p, tname)
}
\arguments{
\item{p}{A pepr::Project object representing your current project.}

\item{tname}{The name of the target to query}
}
\description{
Find path to a target from all your PEP target folders
}
