% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptar.R
\name{peptar_read}
\alias{peptar_read}
\title{Load a target and return it, from all your PEP target folders}
\usage{
peptar_read(p, tname)
}
\arguments{
\item{p}{A pepr::Project object representing your current project.}

\item{tname}{The name of the target to query}
}
\description{
Load a target and return it, from all your PEP target folders
}
