% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_projects.R
\name{source_target_functions}
\alias{source_target_functions}
\title{Helper function to load target-creation functions from PEP config}
\usage{
source_target_functions(config)
}
\arguments{
\item{config}{Path to the PEP config file for the resource targets list.}
}
\description{
This function simple sources the target functions specified in
a 'target_load_functions' attribute of a PEP. It's meant to be
called from a _targets.R file, so you don't have to source those
scripts individually, but can consolidate the configuration into
the PEP config file.
}
