% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_projects.R
\name{tar_build_from_pep}
\alias{tar_build_from_pep}
\title{Target factory for arbitrary functions on PEP samples}
\usage{
tar_build_from_pep(p, tpattern, func, argnames, argvals_pattern, argtypes)
}
\arguments{
\item{p}{PEP.}

\item{tpattern}{Target name, with patterns allowed}

\item{func}{Function to call}

\item{argnames}{List of arg names}

\item{argvals_pattern}{List of arg values, with patterns allowed}

\item{argtypes}{List of arg types (use 'symbol' for targets)}
}
\description{
This target factory creates a target for each element of a PEP. You can
name the targets using a pattern that uses brackets to indicate sample
attributes, like `{sample_name}`.
}
