% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_projects.R
\name{tar_list}
\alias{tar_list}
\title{List the targets available for loading}
\usage{
tar_list(n = 50, fields = c("name", "time", "bytes"), ...)
}
\arguments{
\item{n}{Number of targets to print (with sensible default)}

\item{fields}{Fields passed to tar_meta (with sensible default)}

\item{...}{Any other arguments to pass to tar_meta}
}
\description{
This function is useful for resource projects, where you want to see
what you have available. It's just a simple wrapper that provides
some sensible defaults to \code{tar_meta} for this use case.
In the future I'd like it if this included a text description of the target
}
