% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_projects.R
\name{tar_pep_foreach}
\alias{tar_pep_foreach}
\title{Target factory for arbitrary functions on PEP samples}
\usage{
tar_pep_foreach(
  p,
  tpattern,
  func,
  argvals_pattern,
  argtypes,
  argnames = NULL,
  combine_tname = NULL,
  combine_func_name = "list"
)
}
\arguments{
\item{p}{PEP.}

\item{tpattern}{Target name, with patterns allowed}

\item{func}{Function to call}

\item{argvals_pattern}{List of arg values, with patterns allowed}

\item{argtypes}{List of arg types (use 'symbol' for targets)}

\item{argnames}{List of arg names}

\item{combine_tname}{Target name for a combined target that merges all of the individual targets}

\item{combine_func_name}{Character vector of a function that will combine them.}
}
\description{
This target factory creates a target for each element of a PEP. You can
name the targets using a pattern that uses brackets to indicate sample
attributes, like `{sample_name}`.
}
