% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_projects.R
\name{track_external_target}
\alias{track_external_target}
\title{Experimental target factory to track but not duplicate external files.}
\usage{
track_external_target(tname, ext_tname, func)
}
\description{
EXPERIMENTAL. This function is just a concept that has not been tested.
It will change or disappear in future versions.
}
