% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_projects.R
\name{unitar_list}
\alias{unitar_list}
\title{List all targets available across all projects}
\usage{
unitar_list(n = 50, fields = c("name", "time", "bytes"), tar_dirs = NULL, ...)
}
\arguments{
\item{n}{Number of targets to print (with sensible default)}

\item{fields}{Fields passed to tar_meta (with sensible default)}

\item{tar_dirs}{Target directories to search. NULL will use global option (the default).}

\item{...}{Any other arguments to pass to tar_meta}
}
\description{
If you have available targets from multiple projects, you might be
interested in a quick look at them to see what you have available to load.
This function just wraps tar_meta with a few defaults that are convenient
for this purpose: it increase the default number of targets shown, and
reduces the number of fields. It also runs it across all registered targets
projects.
}
