% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitar.R
\name{unitar_make}
\alias{unitar_make}
\title{A global tar_make function that spans all tar_target folders}
\usage{
unitar_make(tar_folders)
}
\arguments{
\item{tar_folders}{A list of root folders of targets-managed projects}
}
\description{
A global tar_make function that spans all tar_target folders
}
\examples{
tar_folders=""
unitar_meta(tar_folders)
}
