% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitar.R
\name{unitar_meta}
\alias{unitar_meta}
\title{Call tar_meta across multiple tar projects}
\usage{
unitar_meta(tar_dirs, ...)
}
\arguments{
\item{tar_dirs}{A list of root folders of targets-managed projects}

\item{...}{Additional arguments to tar_meta function}
}
\description{
Call tar_meta across multiple tar projects
}
\examples{
tar_dirs=""
unitar_meta(tar_dirs)
}
