% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitar.R
\name{unitar_read}
\alias{unitar_read}
\title{Load targets from a list of possible target repositories}
\usage{
unitar_read(tar_folders, tname)
}
\arguments{
\item{tar_folders}{A priority list of root folders of targets-managed projects}

\item{tname}{The name of the target to query}
}
\description{
This looks in priority order through the list of tprojects, and, if it finds
a matching target, it reads it in and returns it.
}
