% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitar.R
\name{unitar_read_from_path}
\alias{unitar_read_from_path}
\title{Reads a target given a direct path, likely output from \code{unitar_path}}
\usage{
unitar_read_from_path(tpath)
}
\arguments{
\item{tpath}{The path to the target to read}
}
\description{
Reads a target given a direct path, likely output from \code{unitar_path}
}
