% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_projects.R
\name{load_custom_old}
\alias{load_custom_old}
\title{Helper function to build targets from function calls with custom args}
\usage{
load_custom_old(tname, func, argnames, argvals, argtypes)
}
\arguments{
\item{tname}{Target name}

\item{func}{Function to call}

\item{argnames}{List of arg names}

\item{argvals}{List of arg values}

\item{argtypes}{List of arg types (use 'symbol' for targets)}
}
\description{
This function is a target factory that will produce a target for an
arbitrary function call with arbitrary arguments and values.
}
