% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_projects.R
\name{load_external_file}
\alias{load_external_file}
\title{Load targets specified by a CSV}
\usage{
load_external_file(sample_name, rootpath, filepath, func)
}
\arguments{
\item{sample_name}{Name of the target}

\item{rootpath}{Path to where resource files are kept}

\item{filepath}{Relative path to file}

\item{func}{Function to use to load file in R}
}
\description{
This function will allow you to load and track files from a resource folder
}
