% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_projects.R
\name{tar_pep_foreach}
\alias{tar_pep_foreach}
\title{Target factory for arbitrary functions on PEP samples}
\usage{
tar_pep_foreach(
  p,
  tpattern,
  func,
  argnames,
  argvals_pattern,
  argtypes,
  combine_tname = NULL,
  combine_func = list
)
}
\arguments{
\item{p}{PEP.}

\item{tpattern}{Target name, with patterns allowed}

\item{func}{Character string of name of function to call}

\item{argnames}{List of arg names}

\item{argvals_pattern}{List of arg values, with patterns allowed}

\item{argtypes}{List of arg types (use 'symbol' for targets)}

\item{combine_tname}{A target name for a combined}

\item{combine_func}{Character string of name of function for combining samples}
}
\description{
This target factory creates a target for each element of a PEP. You can
name the targets using a pattern that uses brackets to indicate sample
attributes, like `{sample_name}`.
}
