% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitar.R
\name{unitar_exec}
\alias{unitar_exec}
\title{Execute a function across a list of tar_target folders}
\usage{
unitar_exec(tar_dirs = NULL, func = tar_make, ...)
}
\arguments{
\item{tar_dirs}{A list of root folders of targets-managed projects}

\item{func}{Function to execute}

\item{...}{Additional arguments to function}
}
\description{
Execute a function across a list of tar_target folders
}
\examples{
tar_dirs=""
unitar_exec(tar_dirs, tar_meta) 
}
