% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitar.R
\name{unitar_make}
\alias{unitar_make}
\title{A global tar_make function that spans all tar_target folders}
\usage{
unitar_make(tar_dirs = NULL)
}
\arguments{
\item{tar_dirs}{A list of root folders of targets-managed projects}
}
\description{
A global tar_make function that spans all tar_target folders
}
\examples{
tar_dirs=""
unitar_meta(tar_dirs)
}
