% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitar.R
\name{unitar_path}
\alias{unitar_path}
\title{Returns the path to the cache of the given target}
\usage{
unitar_path(tar_folders, tname)
}
\arguments{
\item{tar_folders}{A priority list of root folders of targets-managed projects}

\item{tname}{The name of the target to query}
}
\description{
Returns the path to the cache of the given target
}
