% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitar.R
\name{unitar_read_xprj}
\alias{unitar_read_xprj}
\title{Load targets across projects}
\usage{
unitar_read_xprj(tar_folders, tname)
}
\arguments{
\item{tar_folders}{A priority list of root folders of targets-managed projects}

\item{tname}{The name of the target to query}
}
\description{
This function will allow you to load and track cached targets
from other target projects.
}
